requirements.txt
yfinance
beautifulsoup4
requests
google-genai
supabase
python-dotenv
pandas
